fix: permission by organization and app-error constructor fix

+++ src/modules/events/services/batches/delete-batch.service.ts
@@ -1,14 +1,49 @@
 import { inject, injectable } from 'tsyringe';
 import IBatchRepositoryProvider from '../../infra/orm/repositories/providers/batch-repository.provider';
+import IEventRepositoryProvider from '../../infra/orm/repositories/providers/event-repository.provider';
+import IUserOrganizationRepositoryProvider from '../../../users/infra/orm/repositories/providers/user-organization-repository.provider';
+import UserOrganizationQueryOptions from '../../../users/dtos/user-organization/user-organization-query-options';
+import AppError from '../../../../shared/infra/http/errors/app-error';
 
 @injectable()
 export class DeleteBatchService {
   constructor(
     @inject('BatchRepositoryProvider')
     private batchRepository: IBatchRepositoryProvider,
+    @inject('EventRepositoryProvider')
+    private eventRepository: IEventRepositoryProvider,
+    @inject('UserOrganizationRepositoryProvider')
+    private userOrganizationRepository: IUserOrganizationRepositoryProvider,
   ) {}
 
-  public async execute(batch_id: string) {
+  public async execute(user_id: string, batch_id: string) {
+    const batch = (await this.batchRepository.find({ id: batch_id })).at(0);
+
+    if (!batch) {
+      throw new AppError(404, 'Batch not found.', 'Lote nao encontrado.');
+    }
+
+    const event = (await this.eventRepository.find({ id: batch.event.id })).at(0);
+
+    if (!event) {
+      throw new AppError(404, 'Event not found.', 'Evento nao encontrado.');
+    }
+
+    const userPermissionQueryOptions = {
+      user_id,
+      organization_id: event.organization.id,
+    } as UserOrganizationQueryOptions;
+
+    const userOrganization = (await this.userOrganizationRepository.find(userPermissionQueryOptions)).at(0);
+
+    if (!userOrganization) {
+      throw new AppError(
+        403,
+        'User does not have permission to delete batch in this organization.',
+        'Usuario nao tem permissao para deletar lote nesta organizacao.',
+      );
+    }
+
     const rowsDeleted = await this.batchRepository.delete(batch_id);
     return { message: 'Batch deleted successfully.', deleted: rowsDeleted };
   }
