+++ src/modules/events/services/events/delete-event.service.ts
@@ -1,14 +1,40 @@
 import { inject, injectable } from 'tsyringe';
 import IEventRepositoryProvider from '../../infra/orm/repositories/providers/event-repository.provider';
+import IUserOrganizationRepositoryProvider from '../../../users/infra/orm/repositories/providers/user-organization-repository.provider';
+import UserOrganizationQueryOptions from '../../../users/dtos/user-organization/user-organization-query-options';
+import AppError from '../../../../shared/infra/http/errors/app-error';
 
 @injectable()
 export class DeleteEventService {
   constructor(
     @inject('EventRepositoryProvider')
     private eventRepository: IEventRepositoryProvider,
+    @inject('UserOrganizationRepositoryProvider')
+    private userOrganizationRepository: IUserOrganizationRepositoryProvider,
   ) {}
 
-  public async execute(event_id: string) {
+  public async execute(user_id: string, event_id: string) {
+    const event = (await this.eventRepository.find({ id: event_id })).at(0);
+
+    if (!event) {
+      throw new AppError(404, 'Event not found.', 'Evento nao encontrado.');
+    }
+
+    const userPermissionQueryOptions = {
+      user_id,
+      organization_id: event.organization.id,
+    } as UserOrganizationQueryOptions;
+
+    const userOrganization = (await this.userOrganizationRepository.find(userPermissionQueryOptions)).at(0);
+
+    if (!userOrganization) {
+      throw new AppError(
+        403,
+        'User does not have permission to delete event in this organization.',
+        'Usuario nao tem permissao para deletar evento nesta organizacao.',
+      );
+    }
+
     const rowsDeleted = await this.eventRepository.delete(event_id);
     return { message: 'Event deleted successfully.', deleted: rowsDeleted };
   }
